backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
langchain>=0.3.13
langchain-openai>=0.3.0
langchain-core>=0.3.29
pydantic==2.10.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
numpy==2.2.1
